refactor(search): move useFormValues hook to a new file
LIIKUNTA-567
